+++ iot/src/main/java/de/xavaro/android/iot/things/IOTLocation.java
@@ -15,7 +15,7 @@ public class IOTLocation extends IOTThing
 
     public Double fixedLatFine;
     public Double fixedLonFine;
-    public Float fixedAltFine;
+    public Double fixedAltFine;
 
     public IOTLocation()
     {
